HiQ Board Game
Reffered to by many names, HiQ is a board game often played with marbles
wherein the game begins with an  empty hole on a oard intoch 'pegs' may
jump. In the HiQ programs, the user specifies this initial empty hole
along with the number of holes that should be left on the board once the
game is over. Pegs may jump into an empty hole by jumping over other
pegs. The programs read in a list of valid moves from a text file. The
list of moves have the format (from, over, into). For a move to be legal
at a given point, the from and over holes must contain pegs while the
into whole must be empty.